requirements.txt
absl-py==0.15.0
addict==2.4.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
black==23.1.0
bleach==6.0.0
cachetools==5.3.0
ccimport==0.4.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.2
contourpy==1.0.7
cumm-cu114==0.4.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
fire==0.5.0
flake8==6.0.0
flatbuffers==1.12
fonttools==4.39.2
fqdn==1.5.1
gast==0.3.3
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==2.10.0
idna==3.4
imageio==2.26.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipdb==0.13.13
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lark==1.1.5
llvmlite==0.31.0
lyft-dataset-sdk==0.0.8
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcls==0.25.0
mmcv-full==1.3.13
mmdet==2.14.0
-e git+ssh://git@github.com/Abyssaledge/FSDV2.git@fcb9b7c0c6e4d501bae31f196c6e269376383251#egg=mmdet3d
mmsegmentation==0.30.0
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.7.3
-e git+https://github.com/tencent/ncnn@03550ba532314870464456924891a510eb32e38c#egg=ncnn&subdirectory=python
nest-asyncio==1.5.6
networkx==2.2
ninja==1.11.1
notebook==6.5.3
notebook_shim==0.2.2
numba==0.48.0
numpy==1.19.5
nuscenes-devkit==1.1.10
oauthlib==3.2.2
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
onnxruntime==1.8.1
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pccm==0.4.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.1.1
plotly==5.13.1
pluggy==1.0.0
plyfile==0.8.1
portalocker==2.7.0
prettytable==3.6.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.4
pycocotools==2.0.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.14.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
pytest==7.2.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.1
qtconsole==5.4.1
QtPy==2.3.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
Shapely==1.8.5
six==1.15.0
sniffio==1.3.0
soupsieve==2.4
spconv-cu114==2.3.3
stack-data==0.6.2
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.17.1
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile==2023.3.15
tinycss2==1.2.1
tomli==2.0.1
torch==1.8.0
torch-scatter @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-scatter_1631018401613/work
torchaudio==0.8.0a0+a751e1d
-e git+ssh://git@github.com/Abyssaledge/TorchEx.git@32eabf726169c2c4437428425a6d27400f8efc3a#egg=torchex
torchvision==0.9.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trimesh==2.35.39
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1678559861143/work
uri-template==1.2.0
urllib3==1.26.15
waymo-open-dataset-tf-2-4-0==1.4.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.12.1
yapf==0.32.0
zipp==3.15.0
